requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
beautifulsoup4==4.13.4
python-dotenv==1.1.1
pypdf==5.4.0
streamlit==1.43.2
PyMuPDF==1.26.4
